requirements.txt
streamlit>=1.29.0
openai>=1.0.0
mysql-connector-python>=8.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
altair>=5.0.0
